Example of authelia lite on docker swarm
Just an example docker-compose file for the deployment of Authelia lite on docker swarm. Its just a modified version of the regular docker-compose file and should be compatible with the configuration and users file. Primary changes were removing depreciated options (name, exposed ports, and restart), moving the labels under deploy, enabling swarm mode on Traefik and added a loadbalencer for each available service.