ci(buildkite): utilise conventional-changelog for release notes
Instead of generating our changelog based on crude modifications utilising git log we now utilise conventional-changelog.

conventional-changelog utilises the angular commit structure to categorise and display the changelog for 3 types (fix,feat,perf) and each of the change scopes are identified in the changelog too.

An example of the output for v4.26.0 can be found below:

# [4.26.0](https://github.com/authelia/authelia/compare/v4.25.2...v4.26.0) (2021-02-02)

### Bug Fixes

* **handlers:** refresh user details on all domains ([#1642](https://github.com/authelia/authelia/issues/1642)) ([60ff16b](https://github.com/authelia/authelia/commit/60ff16b518e54bcc2cbb7a3a3441c47b1d382be6))

### Docker Container
* `docker pull authelia/authelia:4.26.0`