fix: traefik was not properly redirected after consent approval.
This is due to the fact that Traefik only forwards scheme and host information
through X-Forwarded-* headers. In this change we get the URI directly from the
webserver instead so there is no needed for this information to be forwarded.

We can probably do the same in other instances where we use X-Forwarded-URI
today.