fix(oidc): adjust kid length per RFC
As per https://datatracker.ietf.org/doc/html/draft-ietf-jose-json-web-key#section-8.1.1 the kid should not exceed 8 chars. As such the kid will now be generated using SHA1 and shortened to the first 7 chars. While it's allowed to exceed 8 chars, it must only be done so with a compelling reason, which we do not have.