Disable access control rule configuration validation
The current requirement of having rules if the default policy is set
to deny blocks any possibility of adding rules only at run time.

Also, the warning that no rules are configures is also disabled as
the warning may be incorrect as rules may be added after the config
file has been validated.