fix(session): change KeyPrefix of Redis provider
Change KeyPrefix of Redis provider by wrapping it in curly braces, so all entries are stored in single slot.
Hashing algorithm of open-source Redis (including Sentinel) does not allow customization, but has one built-in policy: if curly braces are found, string in it is taken as input to hashing function instead if entire key. https://redis.com/blog/redis-clustering-best-practices-with-keys/

This resolves #2323 where `CROSSSLOT` error occurred.