feat(notifier): dedicated sender_name configuration option
This change makes the configuration of the 'sender' slightly different. Instead of specifying the name of the sender in the 'sender' option you must now use the 'sender_name' option. Providing anything other than an email address for the 'sender' is not supported. They will be combined in the format of '{sender_name} <{sender}>'. The reason why this change is necessary is  that the MAIL FROM command format is completely different in input format compared to the format of the FROM header.