feat(handlers): flexible 1FA authorization header
Basic auth is currently supported either with the
"Proxy-Authorization"-header or by using the endpoint
"/api/verify?auth=basic" to use the "Authorization"-header.
Sometimes an application needs the default HTML login page for human
users, and authorization with "Authorization"-header as a token
replacement.

The "auth=both" endpoint enables to first check "Proxy-Authorization",
then "Authorization" headers and redirects to HTML login page if both
are not found.