docs: Clarify port 465 is the preferred port
As per the 2018 RFC 8314, this port is no longer intended for SMTPS, but repurposed as a _submissions service_ port.

Where available, it is advised as the port that should be preferred over any port which uses STARTTLS.

Note that as ports intended only for mail submission (ports 465 & 587), authentication is required to send mail through. Otherwise, use the standard SMTP port 25.