docs: update Seafile OIDC integration
Remove the `groups` scope from the Seafile config since Seafile doesn't
use this information (as far as I can tell) and the given Authelia
config doesn't even grant this scope.
Fix the `OAUTH_ATTRIBUTE_MAP` in the Seafile config. When using the old
values Seafile was reporting "Required user info not found" and login
would fail.
Add a note regarding an issue with the provided configuration and hints
on how to mitigate. Also add a step for installing a required pip
package.