docs: fix copy paste typo (#4815)
Fixed a copy-paste bug in the description of the idle server timeout. This change only fixes that bug, but this part of the documentation need better explaining. (Something called "idle" in the "Server Timeouts" section of course configures server idle timeout, but what is server in this context, when happens when the time defined in idle expires, etc.)