fix(commands): internal services not cleaned up properly
This fixes a race condition which in rare circumstances can cause the process to hang during a shutdown. While unrelated this also adds additional trace logging to the shutdown process to better capture each stage to better facilitate debugging in the future specifically when one particular service is taking time to stop.