feat(templates): add multiline helper functions (#6227)
This adds some multiline helper functions to the templates engine. Specifically 'mquote', 'msquote', and 'mindent' which are helper functions which when presented with multiline strings have similar behaviour to their similarly named counterparts. The 'mindent' function takes an additional input after the indent number which indicates the YAML multiline formatter to add if it's a multiline string.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>