feat(totp): reuse policy
This prevents the reuse of previously used codes being reused within the validity period. This is prevented in a smart way by recording the step multiplied by the period and hashing that value as a string with the HMAC-SHA256 algorithm, with a HMAC key unique to recording these values (auto generated).

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>