fix(handlers): bearer authorization assumed intention (#6920)
This fixes an issue where bearer scheme authorization headers are assumed to be targeted to Authelia. This change ignores bearers which are not potentially Authelia Access Tokens and skips tem as a potential source. This also fixes a situation where unknown or unregistered schemes results in an error when the strategy (not authorization itself) should just be skipped.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>