fix(configuration): missing trailing slash
The trailing slash in the Authelia URL when missing in the path can cause situations where a user can provide an invalid configuration and not be aware which can be rather difficult to diagnose. This fixes that issue by ensuring if there is a path that it ends with a '/'.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>