fix(authorization): mfa not detected in custom policies
This fixes an issue where the MFA enabled state is not correctly detected if the only 'two_factor' policy configured was a custom OpenID Connect 1.0 policy. This ends up being a bad UX as admins have to configure a dummy 'two_factor' policy instead of it being automatically detected.

Fixes #7103.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>