fix(authentication): memberof included in groups searches (#7359)
This fixes an issue where the memberof attribute is requested needlessly. While not necessarily a bug it's not really a feature either. The only reported issue with the current functionality is an annoying log message. In the instance this turns out to be a regression this commit should be immediately reverted.

Closes #7310

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>