fix: busybox wget zombie ssl client (#7498)
This should prevent the issues with the broken busybox wget by installing the musl wget. It should be noted that while this is labelled as a fix this does not fix the actual issue as the issue is an upstream bug with the busybox wget TLS client where it appears to leave zombie processes on any TLS request. This is just a workaround.

Fixes #7463

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>