fix(web): exact dates too verbose for good ux (#7650)
The existing date format for creation time/last used time for 2FA methods in the user settings menu is a bit lengthy and doesn't render too well even on desktop devices. This changes it to a delta-time format. Instead of showing the date, it shows the time since the last event e.g. just now, one day ago, 2 years ago etc. This leads to a better UX.