refactor(suites): prevent log truncation in buildkite (#7967)
The changes in the pull request modify the `Up` method of the `DockerEnvironment` struct in the `internal/suites/docker.go` file. A conditional check for the `CI` environment variable is added, altering the command executed based on its value. If `CI` is set to `t`, the command includes additional flags; otherwise, it remains simpler. Minor formatting adjustments were also made to align the opening brace of the method with its signature.

Signed-off-by: Amir Zarrinkafsh <nightah@me.com>